// frontend/src/types/generated/ProviderRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ProviderRequestKind } from "./ProviderRequestKind.js";
import type { ProviderRequestPrompt } from "./ProviderRequestPrompt.js";

export type ProviderRequest = { provider_request_id: string, request_kind: ProviderRequestKind, title: string, description: string, timeout_seconds: number, prompt: ProviderRequestPrompt, };